refactor: split ReplyItem component into multiple files
